requirements.txt
tqdm 
PyYAML
numpy
opencv-python
pandas
json5
jsonargparse
matplotlib==3.9.3
Pillow>=9.0.1
accelerate>=1.2.0
bitsandbytes>=0.45.0
torch>=2.1.2
torchvision>=0.16.2
sentence-transformers==3.3.1
transformers>=4.37
datasets>=3.1.0
huggingface-hub==0.25.0
diffusers>=0.31.0
peft==0.10.0
pybrisque


